_multiprocessing: extract into module/_multiprocessing/ following the _socket pattern
importing.rs shrinks by 244 lines.
