+++ pyre/pyre-interpreter/src/module/_multiprocessing/mod.rs
@@ -1,0 +1,8 @@
+//! _multiprocessing module — PyPy: pypy/module/_multiprocessing/
+//!
+//! Exposes `SemLock(kind, value, maxvalue, name, unlink)` and
+//! `sem_unlink(name)`.  Single-threaded pyre still needs the methods to
+//! exist so multiprocessing.py teardown survives.
+
+pub mod interp_multiprocessing;
+pub mod moduledef;
